// src/shared/api/generated/types/GlobalHtmlSectionResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type GlobalHtmlSectionResponseDto = {
  /**
   * @type number
   */
  definitionId: number;
  /**
   * @type string
   */
  key: string | null;
  /**
   * @type string
   */
  name: string;
  /**
   * @type number
   */
  revision: number;
  /**
   * @type number
   */
  iframeHeight: number | null;
  /**
   * @type number
   */
  visiblePlacements: number;
  /**
   * @type number
   */
  hiddenPlacements: number;
  /**
   * @type number
   */
  totalPlacements: number;
  /**
   * @type string | undefined
   */
  html?: string;
  /**
   * @type string | undefined
   */
  css?: string;
  /**
   * @type string | undefined
   */
  javascript?: string;
};