// src/shared/api/generated/types/PageRegistrySummaryDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type PageRegistrySummaryDto = {
  /**
   * @type string
   */
  pageKey: string;
  /**
   * @type string
   */
  title: string;
  /**
   * @type string
   */
  routePattern: string;
  /**
   * @type number
   */
  revision: number;
  /**
   * @type number
   */
  totalSections: number;
  /**
   * @type number
   */
  visibleSections: number;
  /**
   * @type number
   */
  hiddenSections: number;
  /**
   * @type number
   */
  pageCustomHtmlSections: number;
  /**
   * @type number
   */
  globalCustomHtmlSections: number;
  /**
   * @type boolean
   */
  isMaterialized: boolean;
};